djangoBackend/pyproject.toml
[tool.poetry]
name = "djangobackend"
version = "0.1.0"
description = ""
authors = ["Christian Q. Cleofas <cleofastiano27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"
psycopg2 = "^2.9.9"
django = "^3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
